% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_likelihood_ratios.R
\name{add_likelihood_ratios}
\alias{add_likelihood_ratios}
\title{Compute the likelihood ratio of pathogenicity}
\usage{
add_likelihood_ratios(df, value = NULL, prior = 0.1, group = NULL)
}
\arguments{
\item{df}{A dataframe. Must contain a \code{class} column with values 'P'
(pathogenic) and 'B' (benign) variants, and a numeric column containing the
variant effect scores. At least 10 occurrences of each class is required.}

\item{value}{(optional) A character string indicating the name of the numeric
column in \code{df} containing the scores. If not provided, likelihood ratios
will be computed for all numeric columns in the dataframe.}

\item{prior}{A scalar in the range 0-1 representing the prior probability
of pathogenicity. Default 0.1.}

\item{group}{(optional) A character string indicating the name of the column
with the grouping variable. Default NULL.}
}
\value{
The input dataframe with new columns for the likelihood ratio, in
the form \code{column_name_lr}, and \code{column_name_lr_lower} and
\code{column_name_upper} for the confidence bounds. The classifications
under the predefined prior are given in \code{column_name_evidence}.
}
\description{
This function calculates the likelihood ratio (equivalent to the odds of
pathogenicity) based on a numeric score, e.g., fitness scores from MAVE
experiments or computational predictors. The input data should have at least
one numeric column with the score of interest, and another column named
\code{class}, containing pathogenic ('P') and benign ('B') truthset labels
for the variants. Likelihood ratios will also be computed for unlabelled
variants or those different from 'P' or 'B' in the data.
}
\details{
The function estimates the likelihood ratio of pathogenicity for each input
score by resampling Gaussian kernel density estimates of the 'P' (pathogenic)
and 'B' (benign) score distributions. Densities are evaluated on a common
grid and interpolated using linear interpolation. To stabilise the ratio in
regions where densities approach zero, a variance-based penalty is computed
from log-density ratios across 1,000 bootstrap replicates. This penalty is
used to regularise the log-likelihood ratio matrix. The log-likelihood ratios
are monotonised in the principal direction of association with the input
scores. Final estimates for each score include the point estimate (median)
and the 95\% confidence interval for the location of the median, derived from
the resamples and mapped from the common grid to the input scores.
}
\examples{
# load the example data provided with the package
library(acmgscaler)
data(variant_data, package = 'acmgscaler')

# compute likelihood ratios
add_likelihood_ratios(
  df = variant_data,
  # provide the column containing the variant scores (optional)
  value = 'score',
  # provide a grouping variable (optional)
  group = 'gene',
  # prior probability of pathogenicity
  prior = 0.1
)


}
\references{
Richards et al., 2015. Modeling the ACMG/AMP variant
classification guidelines as a Bayesian classification framework.
\emph{Genetics in Medicine}.
\href{https://doi.org/10.1038/gim.2017.210}{DOI: 10.1038/gim.2017.210}

Tavtigian et al., 2018. Standards and guidelines for the
interpretation of sequence variants: a joint consensus recommendation of
the American College of Medical Genetics and Genomics and the Association
for Molecular Pathology.
\emph{Genetics in Medicine}.
\href{https://doi.org/10.1038/gim.2015.30}{DOI: 10.1038/gim.2015.30}

Brnich et al., 2019. Recommendations for application of the
functional evidence PS3/BS3 criterion using the ACMG/AMP sequence variant
interpretation framework.
\emph{Genome Medicine}.
\href{https://doi.org/10.1186/s13073-019-0690-2}{DOI: 10.1186/s13073-019-0690-2}

Pejaver et al., 2022. Calibration of computational tools for
missense variant pathogenicity classification and ClinGen recommendations
for PP3/BP4 criteria.
\emph{The American Journal of Human Genetics}.
\href{https://doi.org/10.1016/j.ajhg.2022.10.013}{DOI: 10.1016/j.ajhg.2022.10.013}

van Loggerenberg et al., 2023. Systematically testing human HMBS
missense variants to reveal mechanism and pathogenic variation
\emph{The American Journal of Human Genetics}.
\href{https://doi.org/10.1016/j.ajhg.2023.08.012}{DOI: 10.1016/j.ajhg.2023.08.012}
}
