% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_to_posterior.R
\name{lr_to_posterior}
\alias{lr_to_posterior}
\title{Conversion between likelihood ratio and posterior probability}
\usage{
lr_to_posterior(lr, prior = 0.1)
}
\arguments{
\item{lr}{A numeric vector of likelihood ratios.}

\item{prior}{A scalar in the range 0-1 representing the prior probability
of pathogenicity. Default 0.1.}
}
\value{
A numeric vector of the same length as \code{lr} containing the posterior
probability of pathogenicity values.
}
\description{
#' @description
This function converts positive likelihood ratios into posterior probability
of pathogenicity.
}
\examples{
library(acmgscaler)
lr_to_posterior(lr = c(3.5, 35, 350), prior = 0.1)


}
